feat: Add metadata filtering support to BM25 Retriever and update documentation
